Working prototype of Select
