+++ packages/components/src/Form/Inputs/Select/index.ts
@@ -24,6 +24,6 @@
 
  */
 
-export * from './Select2'
+export * from './Select'
 export * from './SelectContext'
 // export * from './SelectOption'
